Update README.md
Beta Bank customers are leaving: little by little, chipping away every month. The bankers figured out it’s cheaper to save the existing customers rather than to attract new ones. We need to predict whether a customer will leave the bank soon. You have the data on clients’ past behavior and termination of contracts with the bank.Build a model with the maximum possible F1 score. To pass the project, you need an F1score of at least 0.59. Check the F1 for the test set